requirements.txt
google-analytics-data>=0.14.0
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0
flask>=2.0.0
requests>=2.25.0
python-dotenv>=0.19.0
